requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.4
django-toolbelt==0.0.1
gunicorn==19.3.0
psycopg2==2.6.1
static3==0.6.1
wheel==0.24.0
django-registration-redux==1.2
django-mathfilters==0.3.0
django-mailgun==0.8.0
